SynapseAI rewrite/requirements.txt
psutil>=5.9.0
flask>=2.3.0
flask-cors>=4.0.0
pywebview>=4.4
lupa>=2.0
